Deleted Unused Synthesizers
