--- cinnamon-synthetization/synthetic_tabular_data_generator/synthesizer_config/llm_nearest_neighbor_knowledge_grounded_text_synthesis.yaml
@@ -1,79 +1,0 @@
-name: llm_nearest_neighbor_knowledge_grounded_text_synthesis
-version: "0.1"
-type: cross-sectional
-display_name: LLM Knowledge-grounded Text Synthesis
-description: Generate synthetic medical free-text fields in two sequential LLM steps. In the first step, the model optionally repairs inconsistent non-TEXT values using nearest-neighbor original rows as structured references. In the second step, it generates the free-text fields from the repaired row using nearest-neighbor reference texts from the original data as semantic and stylistic guidance. Knowledge-base grounding is not implemented yet and is currently exposed only as NOT_IMPLEMENTED.
-URL: /start_synthetization_process/llm_nearest_neighbor_knowledge_grounded_text_synthesis
-processing_capabilities:
-  data_modality: mixed
-  generation_scope: text_only
-configurations:
-  llm_profile:
-    display_name: LLM Profile
-    description: Select which configured LLM instance should be used.
-    parameters:
-      - name: llm_profile
-        type: string
-        label: LLM Profile
-        description: Select the configured LLM profile (model endpoint and credentials).
-        default_value: ""
-        values: []
-        mandatory: true
-  model_parameter:
-    display_name: Similarity and Knowledge Parameters
-    description: Configure how nearest-neighbor examples are selected. Knowledge bases are not implemented yet.
-    parameters:
-      - name: few_shot_rows
-        type: integer
-        label: Example Rows
-        description: Number of nearest-neighbor original rows to include as in-context examples for each synthetic row.
-        default_value: 5
-        min_value: 0
-        max_value: 200
-      - name: similarity_strategy
-        type: string
-        label: Similarity Strategy
-        description: Strategy used to identify the most similar original rows for each synthetic row. Attributes retrieves nearest neighbors using only non-TEXT attributes. Random samples original rows uniformly.
-        default_value: Attributes
-        values: ["Attributes", "Random"]
-      - name: knowledge_source_type
-        type: string
-        label: Knowledge Source Type
-        description: Knowledge-base support is not implemented yet.
-        default_value: NOT_IMPLEMENTED
-        values: ["NOT_IMPLEMENTED"]
-  model_fitting:
-    display_name: Prompt Configuration
-    description: Configure domain context and whether structured values may be minimally corrected for consistency.
-    parameters:
-      - name: user_prompt_domain_context
-        type: string
-        multiline: true
-        rows: 5
-        label: Domain Context
-        description: Optional domain context to guide realistic text generation, for example clinical notes, registry data, or claims narratives.
-        default_value: ""
-        mandatory: false
-      - name: allow_structured_corrections
-        type: boolean
-        label: Allow Structured Corrections
-        description: If true, the model may minimally adjust non-TEXT values when combinations are implausible.
-        default_value: true
-  sampling:
-    display_name: Sampling
-    description: Configure LLM decoding behavior for nearest-neighbor and knowledge-grounded text generation.
-    parameters:
-      - name: temperature
-        type: float
-        label: Temperature
-        description: Controls randomness of token sampling. Lower values are more deterministic.
-        default_value: 0.3
-        min_value: 0.0
-        max_value: 2.0
-      - name: top_p
-        type: float
-        label: Top-p
-        description: Controls nucleus sampling by limiting candidate tokens to a probability mass.
-        default_value: 0.9
-        min_value: 0.0
-        max_value: 1.0
