--- cinnamon-synthetization/synthetic_tabular_data_generator/synthesizer_config/llm_text_only_semantic_variation_synthesis.yaml
@@ -1,64 +1,0 @@
-name: llm_text_only_semantic_variation_synthesis
-version: "0.1"
-type: cross-sectional
-display_name: LLM Text-only Semantic Variation Synthesis
-description: Generate a new single TEXT column for fictional but plausible patients by sampling source texts and asking an LLM to create semantically similar variations. The generated free text may add, omit, or change details as long as it remains close in topic and style to the sampled source.
-URL: /start_synthetization_process/llm_text_only_semantic_variation_synthesis
-processing_capabilities:
-  data_modality: text_only
-  generation_scope: text_only
-configurations:
-  llm_profile:
-    display_name: LLM Profile
-    description: Select which configured LLM instance should be used.
-    parameters:
-      - name: llm_profile
-        type: string
-        label: LLM Profile
-        description: Select the configured LLM profile.
-        default_value: ""
-        values: []
-        mandatory: true
-  model_fitting:
-    display_name: Prompt Configuration
-    description: Configure optional domain context for semantically related fictional text generation.
-    parameters:
-      - name: required_attributes
-        type: named_list
-        label: Required Attributes
-        description: Define named concepts that must be explicitly mentioned in the generated text. Each entry can add an optional description or constraint for the generator.
-        default_value: []
-        mandatory: false
-      - name: user_prompt_domain_context
-        type: string
-        multiline: true
-        rows: 5
-        label: Domain Context
-        description: Optional domain context to guide terminology, style, and plausibility during generation.
-        default_value: ""
-        mandatory: false
-  sampling:
-    display_name: Sampling
-    description: Configure how many synthetic text rows to generate and how deterministic the LLM output should be.
-    parameters:
-      - name: num_samples
-        type: integer
-        label: Number of Samples
-        description: Number of synthetic text rows to generate. Source rows are selected randomly and may be reused when needed.
-        default_value: $dataset.original.numberRows
-        min_value: 1
-        max_value: 1000000
-      - name: temperature
-        type: float
-        label: Temperature
-        description: Controls randomness of token sampling. Higher values encourage more varied generations.
-        default_value: 0.8
-        min_value: 0.0
-        max_value: 2.0
-      - name: top_p
-        type: float
-        label: Top-p
-        description: Controls nucleus sampling by limiting candidate tokens to a probability mass.
-        default_value: 0.95
-        min_value: 0.0
-        max_value: 1.0
